This is the first commit
I've already done a lot but havent created a github repo for storage.
This is just a back up of it.  These are the things that I have done so
far:

1) Read in the provided .mat file into python
2) Did some preliminary visualization of the slices just to see what we
are dealing with
3) Did some preliminary 3D volumetric to mesh generation using marching
cube algorithm.

Future work:
This is basically working with Valentin to do some segementation to
clean up the images.