Generate Stack Of Images for 3D Slicer
The main objective was to perform some simple thresholding on the slices
and create an image stack of them.  This was then imported in the 3D
slicer and create a volume.  From this volume we can create a mesh.
Finally from this mesh we can import it to mesh lab and do some furhter
cleaning.