Filtering and STL Generation
I have applied the morphological filters and median filters to remove as
much of the small noise in the image as possible.  The veins are
retained. The variables have been saved.  Need to go back and perform
thresholding again though.

There are two main possibilities going forward to generate the STL
files.  The first is the same as before which is going through 3D slicer
and the second will be VTK.

In the long run VTK is ideal for automation so this must be explored
further.