Working Marching Cube Smoothing
I think there are two methods for smoothing the data.

1) Interpolatin. Basically because we don't have enough samples between
the slices it looks very choppy. I think this is inherent in all
scanning type biomedical imaging.

2) Applying a VTK function called vtkWindowedSincPolyDataFilter on it. I
guess for this we can use these examples to try and see which ones works
best.

I have completed the second one and seems to be the fasted route.