Segmentation Attempt
This is an attempt to segment the mouse brain dataset.

I was able to open .mnc files which contain volumetric data and the
supplimentary label mnc file.  I was able to load both but there is a
problem.  We don't know the labeling of the mnc file.  Worst case
scenario which i have done is create a histogram distribution and create
a filter on this to create a mask and then do segmentation that way.