First Attempt Monkey Brain
Working on the Simple ITK and VTK implemenation of the connected
threshold algorithm for segmentation.

I used the one Adriana found using the monkey brain.  There are more
that she found but for the purpose of my work I will stick to this one
for now.

The next step is to perform this segmentation for each slice going
forward.