Human Brain Read and Image
Generated an anaylsis template that I can modify as I go to include the
many functions that I have used so far in the project.

Using this template I did a cleaner version of reading and viewing the
images of MR images.

Next step would be to import them in 3D slicer and perform segmentation.