Image Stack for Gabriel
Michael wanted me to give Gabriel an image stack of the monkey, rat and
human brain with the skull still in tact.  This was the session I did it
on.  I also uploaded these in a folder in the one drive folder under the
image stack folder and in a folder called forGabriel